% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_manipulation.R
\name{merge_years}
\alias{merge_years}
\title{Combine weather data from different years into a tlb, and write it
out as a CSV file (optional).}
\usage{
merge_years(years, write_file = FALSE, filename = "all_data.csv")
}
\arguments{
\item{years}{A numeric array}

\item{write_file}{A boolean to determine whether to write the merged data
out to a file}

\item{filename}{A text string specifying the filename to write the data
structure out to}
}
\value{
A data frame containing weather data from the specified years.
(Also writes out a csv of the data - see write_file and filename
parameters above)
}
\description{
Combine weather data from different years into a tlb, and write it
out as a CSV file (optional).
}
\examples{
years <- c(2018,2019)
merge_years(years, TRUE, "my_data.csv")
merge_years(years)
}
