% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_visualisation.R
\name{plot_windrose}
\alias{plot_windrose}
\title{Wind rose plot}
\usage{
plot_windrose(data, speed, direction, spdres = 2, dirres = 30,
  spdmin = 2, spdmax = 20, spdseq = NULL, palette = "YlGnBu",
  countmax = NA, printplot = TRUE)
}
\arguments{
\item{data}{A data frame containing wind speed and direction in degrees}

\item{speed}{The name of the column in the data frame containing wind speed}

\item{direction}{The name of the column in the data frame containing wind direction}

\item{spdres}{Resolution of wind speed bins}

\item{dirres}{Resolution of wind direction bins}

\item{spdmin}{Maximum wind speed to use in binning}

\item{spdmax}{Minimum wind speed to use in binning}

\item{palette}{Palette to use in plotting}

\item{countmax}{Maximum axis limit for plotting}
}
\value{
A ggplot2 object containing the windrose plot
}
\description{
Plot a wind rose diagram, showing the relative frequency of winds coming
from each point of the compass
}
\details{
Radar plot code modified from answers to this Stackoverflow question: https://stackoverflow.com/questions/9614433/creating-radar-chart-a-k-a-star-plot-spider-plot-using-ggplot2-in-r/10820387
}
\examples{
# Assuming EasyWeather data files called 2018.csv and 2019.csv are in local
# directory:
years <- c(2018,2019)
test_data <- merge_years(years)
p <- plot_windrose(test_data, "wind_km_h", "wind_direction_deg", dirres = 15)
}
